Output of TZ-grid + solved problem TZ --> percentage as double now

+++ runparameter.h
@@ -37,7 +37,7 @@ public:
     int nb_LU;
     //! width of transition zone
     int TZ_width;
-    int TZ_percentage;
+    double TZ_percentage;
     static std::string size_order;
 
     //! disturbances
